splite commands
