+++ modules/landing_zone/main.tf
@@ -1,10 +1,39 @@
+resource "null_resource" "landing_zone_config" {
+  triggers = {
+    command = "${var.landing_zone_command}"
+    components = "${md5(jsonencode(var.landing_zone_components))}"
+    region = "${var.region}"
+    account_id = "${var.account_id}"
+  }
+
+  "provisioner" "local-exec" {
+    when    = "create"
+    command = "sh ${path.module}/scripts/add_config.sh"
+    environment = {
+      ROOT_PATH = "${var.root_path}"
+      COMMAND = "${var.landing_zone_command}"
+      COMPONENTS = "${jsonencode(var.landing_zone_components)}"
+      REGION = "${var.region}"
+      ACCOUNT_ID = "${var.account_id}"
+    }
+  }
+
+  provisioner "local-exec" {
+    when    = "destroy"
+    command = "python ${path.module}/scripts/sub_config.py"
+    environment = {
+      root = "${var.root_path}"
+      components = "${jsonencode(var.landing_zone_components)}"
+    }
+  }
+}
+
 resource "null_resource" "landing_zone_apply" {
+  depends_on = ["null_resource.landing_zone_config"]
   triggers = {
     command = "${var.landing_zone_command}"
     components = "${md5(jsonencode(var.landing_zone_components))}"
     timestamp = "${timestamp()}"
-    region = "${var.region}"
-    account_id = "${var.account_id}"
   }
 
   "provisioner" "local-exec" {
@@ -14,8 +43,6 @@ resource "null_resource" "landing_zone_apply" {
       ROOT_PATH = "${var.root_path}"
       COMMAND = "${var.landing_zone_command}"
       COMPONENTS = "${jsonencode(var.landing_zone_components)}"
-      REGION = "${var.region}"
-      ACCOUNT_ID = "${var.account_id}"
     }
   }
