+++ modules/landing_zone/scripts/apply.py
@@ -5,14 +5,10 @@ import subprocess
 def main():
     components = eval(os.environ['components'])
     processes = []
-    processes.append(['terrahub', 'configure', '-c', 'template.tfvars.region=' + os.environ['region']])
-    processes.append(['terrahub', 'configure', '-c', 'template.tfvars.account_id=' + os.environ['account_id']])
-
+    
     include = []
     for (k, v) in components.items():
         include.append(k)
-        execWithoutErrors(['terrahub', 'configure', '-i', k, '-c', 'terraform', '-D', '-y'])
-        processes.append(['terrahub', 'configure', '-i', k, '-c', 'terraform.varFile[0]=' + str(v)])
     includeStr = ','.join(include)
     processes.append(['terrahub', 'init', '-i', includeStr])
     processes.append(['terrahub', os.environ['command'], '-i', includeStr, '-y'])
@@ -64,13 +60,6 @@ def execWithErrors(args_list):
         if p.wait() != 0:
             print("Error: failed to execute command:")
             raise Exception(error)
-
-def execWithoutErrors(args_list):
-    p = subprocess.Popen(args_list, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=os.environ['root'])
-    (result, error) = p.communicate()
-    if p.wait() != 0:
-        print("Error: failed to execute command:")
-        print(error)
 
 if __name__ == '__main__':
     RESP = main()
